requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.2
pillow==10.4.0
sqlparse==0.5.1
